Improved handling of multiline spans and spaces.
We now ignore the <autosplit> parameter passed to extract_add_char() and always
split multiline spans. This ensures that spaces are inserted between lines in
the original pdf layout when we join the resulting lines.

Have also made small modification to line-joining to avoid inserting a space if
the first line already ends with a space.

Updated expected output to match new behaviour.