Fixed incorrect vertical splitting of tables.
We now use horizontal lines as well as vertical lines when looking for vertical
contiguous regions corresponding to a single table.

This improves output for electoral_roll.pdf - now outputs single table instead
of multiple separate tables.

Unfortunately this breaks our handling of twotables_2.pdf because it as white
horizontal lines that, if interpreted as table lines, causes table to include
non-table text. Have added a cheat which ignores white horizontal lines in the
new code.